requirements.txt
# Enhanced BTIA-AD Net for Medical VQA
# PyTorch 2.1+ with CUDA 12.x support for RTX 5070 Ti

# Core ML
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers and Model Hub
transformers>=4.36.0
huggingface-hub>=0.20.0
open_clip_torch>=2.24.0
timm>=0.9.12

# Training Utilities
accelerate>=0.25.0
datasets>=2.16.0
safetensors>=0.4.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0.0
omegaconf>=2.3.0

# Logging and Monitoring
tensorboard>=2.15.0
tqdm>=4.66.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
einops>=0.7.0
requests>=2.31.0
